requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
scikit-learn==0.24.2
numpy==1.21.2
pandas==1.3.3
python-docx==0.8.11
PyPDF2==2.0.1
spacy==3.1.3 